// src/app/model/consentimiento.ts
/**
 * Proyecto TDS Backend
 * Descripción
 *
 * OpenAPI spec version: 2.0
 * Contact: info@tecazuay.edu.ec
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Consentimiento { 
    cama?: string;
    canton?: string;
    codUd?: string;
    estado?: string;
    fecha?: Date;
    idConsentimiento?: number;
    instutucionSistema?: string;
    numeroHistoriaClinica?: number;
    parroquia?: string;
    provincia?: string;
    sala?: string;
    servicio?: string;
    unidadOperativa?: string;
}