// src/app/model/informacionTratamientoDTO.ts
/**
 * Proyecto TDS Backend
 * Descripción
 *
 * OpenAPI spec version: 2.0
 * Contact: info@tecazuay.edu.ec
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface InformacionTratamientoDTO { 
    codigo?: number;
    especialidad?: string;
    idConsentimientoFK?: number;
    idInfTrat?: number;
    nombreProfesionalTrat?: string;
    procedimientoPropuesto?: string;
    proposito?: string;
    resultadoEsperado?: string;
    riesgoComplicaciones?: string;
    telefono?: number;
}